:art: Add enhancement in code
Issue: #4
