+++ collector_test.go
@@ -3,6 +3,7 @@ package redigosrv
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"github.com/gomodule/redigo/redis"
 	. "github.com/onsi/ginkgo"
@@ -10,6 +11,15 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	dto "github.com/prometheus/client_model/go"
 )
+
+type poolStatsFake struct{}
+
+func (psf *poolStatsFake) Stats() redis.PoolStats {
+	return redis.PoolStats{
+		ActiveCount: 10,
+		IdleCount:   7,
+	}
+}
 
 var _ = Describe("RedigoCollector", func() {
 	It("should count total of data send using method Publish", func(done Done) {
@@ -316,6 +326,48 @@ var _ = Describe("RedigoCollector", func() {
 		}).Write(&metric)).To(BeNil())
 
 		Expect(metric.GetCounter().GetValue()).To(Equal(1.0))
+	})
+
+	It("should generate description name", func() {
+		collector := NewRedigoCollector(nil, RedigoCollectorDefaultOptions())
+		ch := make(chan *prometheus.Desc)
+		go func() {
+			collector.Describe(ch)
+			close(ch)
+		}()
+		Expect((<-ch).String()).To(ContainSubstring("redigo_active_count"))
+		Expect((<-ch).String()).To(ContainSubstring("redigo_idle_count"))
+	})
+
+	It("should generate custom description name", func() {
+		customName := "mabel"
+		collector := NewRedigoCollector(nil, RedigoCollectorOptions{
+			Prefix: customName,
+		})
+		ch := make(chan *prometheus.Desc)
+		go func() {
+			collector.Describe(ch)
+			close(ch)
+		}()
+		Expect((<-ch).String()).To(ContainSubstring(fmt.Sprintf("redigo_%s_active_count", customName)))
+		Expect((<-ch).String()).To(ContainSubstring(fmt.Sprintf("redigo_%s_idle_count", customName)))
+	})
+
+	It("should test default values", func() {
+		fakePool := poolStatsFake{}
+		collector := NewRedigoCollector(&fakePool, RedigoCollectorDefaultOptions())
+		ch := make(chan prometheus.Metric)
+
+		go func() {
+			collector.Collect(ch)
+			close(ch)
+		}()
+		var metric dto.Metric
+
+		Expect((<-ch).Write(&metric)).To(Succeed())
+		Expect(metric.GetGauge().GetValue()).To(Equal(float64(10)))
+		Expect((<-ch).Write(&metric)).To(Succeed())
+		Expect(metric.GetGauge().GetValue()).To(Equal(float64(7)))
 	})
 
 })
